# drivers/imu: Invensense & Bosch remove transfer perf counters