--- a/src/drivers/imu/invensense/icm20948/ICM20948_AK09916.hpp
+++ b/src/drivers/imu/invensense/icm20948/ICM20948_AK09916.hpp
@@ -88,7 +88,6 @@ private:
 
 	PX4Magnetometer _px4_mag;
 
-	perf_counter_t _transfer_perf{perf_alloc(PC_ELAPSED, MODULE_NAME"_ak09916: transfer")};
 	perf_counter_t _bad_transfer_perf{perf_alloc(PC_COUNT, MODULE_NAME"_ak09916: bad transfer")};
 	perf_counter_t _magnetic_sensor_overflow_perf{perf_alloc(PC_COUNT, MODULE_NAME"_ak09916: magnetic sensor overflow")};