- we already have nearly equivalent data in every sensor publication with the (publication) timestamp and timestamp_sample

Individually perf counters are pretty cheap (2-3 us per call), but it still adds up with multiple sensors at high rates. Looking at the difference between publication timestamp and timestamp_sample is sufficient.

This saves 0.5 - 1% cpu depending on the board (number of IMUs) and IMU_GYRO_RATEMAX.

``` Console
nsh> listener sensor_gyro

TOPIC: sensor_gyro 3 instances

Instance 0:
 sensor_gyro_s
        timestamp: 264937625  (0.003857 seconds ago)
        timestamp_sample: 264937525  (100 us before timestamp)
        device_id: 3670026 (Type: 0x38, SPI:1 (0x00)) 
        x: 0.0090
        y: -0.0045
        z: 0.0143
        temperature: 40.8629
        error_count: 14

Instance 1:
 sensor_gyro_s
        timestamp: 264941837  (0.003741 seconds ago)
        timestamp_sample: 264941741  (96 us before timestamp)
        device_id: 3932170 (Type: 0x3C, SPI:1 (0x00)) 
        x: 0.0050
        y: 0.0029
        z: -0.0046
        temperature: 40.3244
        error_count: 0

Instance 2:
 sensor_gyro_s
        timestamp: 264943743  (0.004815 seconds ago)
        timestamp_sample: 264943711  (32 us before timestamp)
        device_id: 4325386 (Type: 0x42, SPI:1 (0x00)) 
        x: -0.0149
        y: -0.0287
        z: 0.0202
        temperature: nan
        error_count: 1
```